adds verifying capability to the verify service

+++ CodeChallenge/CodeChallenge.DisbursementsVerifier.Models/ProcessedDisbursementData.cs
@@ -5,5 +5,5 @@ public class ProcessedDisbursementData
     public int EmployeeCode { get; set; }
     public int Year { get; set; }
     public int Quarter { get; set; }
-    public int Disbursement { get; set; }
+    public decimal Disbursement { get; set; }
 }
